requirements.txt
discord.py[voice]==2.4.0
python-dotenv==1.0.1
yt-dlp==2025.2.19
PyNaCl==1.5.0
